further work in progress

+++ src/main/java/net/pms/util/jna/StringByReference.java
@@ -1,3 +1,21 @@
+/*
+ * Universal Media Server, for streaming any media to DLNA compatible renderers
+ * based on the http://www.ps3mediaserver.org. Copyright (C) 2012 UMS
+ * developers.
+ *
+ * This program is a free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the Free
+ * Software Foundation; version 2 of the License only.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+ * details.
+ *
+ * You should have received a copy of the GNU General Public License along with
+ * this program; if not, write to the Free Software Foundation, Inc., 51
+ * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ */
 package net.pms.util.jna;
 
 import java.io.UnsupportedEncodingException;
@@ -22,7 +40,7 @@ public class StringByReference extends PointerType {
 	 * @param dataSize the size to allocate in bytes excluding the {@code null}
 	 *            terminator.
 	 */
-	public StringByReference(int dataSize) {
+	public StringByReference(long dataSize) {
 		super(dataSize < 1 ? Pointer.NULL : new Memory(dataSize + 1));
 	}
